feat: Add MetaOr functions for all variables
The respective MetaOr functions can be used so that when a value wasn't
set during the build process, a default value is returned.

For example, `meta.Version()` would have returned an empty string -
using `meta.Version("v1.0.0-dev")` it would return the given string,
unless a version was set during the build process.